Refactor WineTest and update JDK version in misc.xml
Refactored WineTest class to simplify the test setup and added more specific test cases for the Wine class getter and setter methods. The goal was to improve the granularity of the unit tests for more precise testing of individual Wine class properties. The JDK version in the misc.xml was updated to match the actual version being used in the project.