Refactor and extend WineTest unit test
The WineTest class was refactored in this commit to use @BeforeEach annotation for code reuse in every test, enhancing it with extra unit tests for each getter and setter method in the Wine class. Now we also test against default values and test all CRUD operations for each field in Wine. This is to ensure the class works correctly in all expected use cases.