Refactor unit tests in AssortmentTest class
These changes clarify and simplify the code structure in AssortmentTest class. Renamed variables for better readability. Changed Imported List to Collections and Arrays. Removed unnecessary comments. Improved the setUp method by standardizing the wine objects and clearly initializing the Assortment object. Each test was also improved and specific scenarios were added for wine already in/not in the assortment. Added a test for getPrice function to verify price calculation of the wines in the assortment.