"Add JetBrains annotations and improve code documentation"
This commit includes the addition of JetBrains' annotation library to the project's dependencies, which is reflected in an update to the gradle configuration file. It makes use of these annotations for nullability-related purposes, providing extra layer of safety to the code.

In addition, significant updates were made to the code's documentation—enhancing readability and providing more comprehensive information. This includes adding and modifying descriptions for classes and their methods, parameter descriptions, return type descriptions, and more. These changes span across multiple classes including Wine, Assortment, Color, PostgreSqlManager, etc. This helps in better code understanding and increases ease of use when working with these classes and methods.