Update package and module info documentation style
This commit modifies the documentation style in package-info.java files and module-info.java to enhance readability. It replaces the paragraph breaks, indicated by < p >, with bullet points, marked by < ul > and < li >, in order to clearly list what each package or module includes. This makes it easier to quickly identify and understand the functionalities of each package in 'winestock' application.