Remove Gradle publish workflow file and update project version
Deleted the Gradle publish workflow file (.github/workflows/gradle-publish.yml) as it is no longer needed for the project's build pipeline. It is better to manually handle the project's packaging and release process. Also, updated the version in the build.gradle file from "0.1-SNAPSHOT" to "0.1" to indicate the readiness of this version for production.