Refactor wine and assortment classes and update the UI
Refactored the Wine class to make builder fields final. Changed the assortment data model in Assortment class. The year and totalPrice properties were made non-final and a wineNames property was added to store the names of all wines in the assortment. The collection-based constructor was removed, and the add() and remove() methods were updated to adjust the year, totalPrice, and wineNames properties accordingly. Retrieval method for assortments was added in PostgreSQLManager. Updated the UI to reflect these updates.

These updates were necessary to improve data consistency, simplifying the retrieval of assortment data from the database and displaying this data on the UI.