Update JavaFX version and remove unused 'loginButton'
The JavaFX version in WineStock.fxml was updated from 19 to 17.0.6 to match with the current project dependencies. Additionally, 'loginButton' declared in WineStockController.java was removed as it was unused and no reference was found in the rest of the codebase. Both these changes provide cleaner code and efficient dependency management.