Added application icon and improved UI layout
An ImageView was imported in WineStock.fxml and an icon added to the 'Connection' tab alongside tweaks to the existing layout for better visuals. Also, the application's dock and/or taskbar icon is now set on startup, with special handling for macOS. This provides more visual cues to users about the application itself. Renamed the launcher script from 'app' to 'WineStock' in build.gradle for better recognition in user's system. Initialized the Application icon in the WineStockController.