Corrected typos and enhanced documentation
Several typos in the wine bottle size enumeration were fixed. This included the correct spelling of "Jeroboam" and the addition of an accent in "Réhoboam."

The README file was updated to correct the formatting for the command to run the application.

Lastly, the Javadoc for the method returning the list of wines in assortment was enhanced to increase clarity and readability. This was done by adding a line break for separate points.