Add CI/CD for publishing package to Maven and GitHub Packages
This commit introduces a new workflow file for GitHub Actions. The workflow gets triggered each time a release is created. It sets up the environment to run on latest Ubuntu with Java 17 configured. After validating the Gradle wrapper, it runs the 'publish' Gradle task. The username and password for publishing to Maven Central are retrieved from the repository's secrets. In addition, changes in build.gradle file have been made to include 'OSSRH' as a repository in the publishing task, with credentials retrieved from environment variables. These changes are needed to automate the process of publishing artifacts to Maven Central and GitHub Packages.