Refine build.gradle comments and structure for clarity
This commit changes comments and restructures statements in build.gradle file for a better understanding of the code. Original comments were rephrased to be more descriptive, and some statements were moved for clearer structure. Removed jlinkZip task as it was redundant. Included a section to create tasks for generating Javadoc and source Jar. The publishing task is now configured to publish to OSSRH or Sonatype instead of GitHub package repository. This was done to establish a more organized and concise build configuration.