Refactor Wine class and clean up BottleSize class
Wine.java was modified to enhance error handling - instead of converting the color to upper case, we first assign the comment. In case of InvalidBottleVolumeException the comment is set to warn for deletion. The color assignment is done afterwards. This helps in easier identification of invalid wine entries.

In BottleSize.java, a redundant import statement was removed to clean up the code.