Remove InvalidWineColorException, add year validity check
A check for year validity has been added to ensure the year of the wine is not in the future. This led to the removal of InvalidWineColorException.java since wine color validation is now directly done in Wine.java by catching IllegalArgumentException. Also, the project version in build.gradle has been updated. Validation of wine data was made strict and moved to class constructors for better data consistency.