Improve exception handling in unit tests
The previous test cases in WineTest, BottleSizeTest, and AssortmentTest were modified to handle exceptions more appropriately. For instance, in WineTest, instead of just setting the year, the new changes introduce checking if the operation runs without throwing an exception. Similarly, in BottleSizeTest, the test is checking for an InvalidBottleVolumeException if a non-standard value is used. The AssortmentTest was also modified to ensure that the removed wine operation does not throw an exception. These changes were done to make sure our tests catch erroneous behavior during the execution of the test cases.