Update WineStock.fxml file
This commit updates WineStock.fxml file in two areas. It changes the tab alignment in the TabPane by combining minHeight and minWidth attributes into one line. This change improves readability of the code. It also updates the JavaFX version from 17.0.6 to 17.0.7 to keep up with the latest version, and updates the WineStock version text displayed in the application from 0.5 to 0.6 reflecting the new updates and improvements made to the application.