Update version in titles and refine javadoc comments
This change updates the API version in HTML titles from 0.5 to 0.6 across various documentation files. Additionally, the javadoc comments for certain methods in "Wine" and "Wine.Builder" classes are refined to provide a more accurate and detailed description of their behaviors. For instance, the 'setVolume()' is now described as setting the volume of the bottle rather than just setting volume. For the "Builder" class constructors, an emphasis is added on the validation of parameters' values. Also, exceptions that can be thrown by certain methods are mentioned. The changes are aligned to offer clearer and more useful information to users.