Remove unused constant and refine exception handling in Wine class; update .gitignore
A constant `INVALID_WINE_MESSAGE` from the `Wine` class was removed since it was not being used. We've replaced the usage of this constant with a direct throw of an `IllegalArgumentException` while handling `InvalidBottleVolumeException`, providing more clarity to users on the real issue encountered during the volume setting. Also, clean up the .gitignore to ignore the entire .idea directory rather than individual files for a cleaner listing.