Remove XML declaration in misc.xml
XML declaration line in .idea/misc.xml file was removed as it's not required in XML files unless the document declaration differs from the default (UTF-8 encoding, standalone no). The removal of the said line should not affect the application behavior in any which way. The project continues to use UTF-8 encoding, which is the default.