Refactor WineStock.fxml to enhance the readability
In this commit, unnecessary lines in WineStock.fxml are removed and some XML elements are rearranged to improve the readability and maintainability of the code. Moreover, the formatting is streamlined by adjusting the white spaces in the TextField and Button components. The software version is also updated from 0.6 to 0.6.1. These changes do not affect the functioning of the application but simplify the process of understanding and editing the code