Update test case for Wine class ID field
The test case for retrieving the ID field in the Wine class has been updated. The expected result was changed from 5 to 0 to reflect the new expected behaviour of the getId method, after the introduction of a unique ID field for Wine objects in the system for better identification and management.