Refactor error messages to be constant fields
A constant string field, YEAR_AFTER_CURRENT_YEAR_INVALID, was created to hold the error message for any creation or modification of a Wine instance with a year value exceeding the current year.
This refactoring was done to standardize error messages in the Wine class and minimize repetition. It also makes any future changes to the error message easier to implement since it only needs to be modified in one place.
Also, logging statements were updated to use placeholders('{}'), which helps to improve log performance by delaying string creation until necessary.