Upsert WineStock's id in PostgreSQLManager
Refactor the id upsertion of WineStock in PostgreSQLManager.java to make it more readable. Previously, the id was fetched from the ResultSet and directly returned from the method. The current change assigns it to the 'id' variable and also sets the 'id' of the corresponding 'wine' object before returning the id. This way helps clarify that the id fetched from the database doesn't just get returned but is also set in the wine object.