Add ID column to Wine and Assortment tables
Added 'id' attribute to both Assortment and Wine classes and modified their constructors accordingly. Updated TableView in the WineStock.fxml to include a new column for 'id'. Also, enhanced the process of fetching 'id' from ResultSet in PostgreSQLManager for better code readability. The changes are made to provide unique identifiers for elements in tables and improve code management.