Add default constructor to Assortment class
This commit introduces the default constructor for Assortment class. It creates an empty Assortment instance with id set to -1, year set to null, totalPrice set to 0, and wineNames initialized as an empty string. This change simplifies the creation of new Assortment instances in scenarios where the initial state is preferable or where the methods for setting individual fields are to be used subsequently. It also simplifies changes in test class handling.