Upgrade Gradle action, downgrade upload-release-asset in build workflow
The Gradle build action on our GitHub workflow was updated from v1 to v2 to take advantage of the latest features and fixes. However, the upload-release-asset action was downgraded from v2 to v1 due to compatibility issues faced with the v2. Changes are reflected in .github/workflows/release.yml file.