Update GitHub token variable in workflow file
Updated the GitHub token variable in GitHub Actions workflow file from `GITHUB_TOKEN` to `MY_GITHUB_TOKEN`. This was necessary as we replaced the standard GitHub token with a custom one for increased security and permissions control. The workflow needs this token to upload the built JAR during the release process.