Update the Gradle build file to support modular Java
To support the Java 11+ module system, a 'Main-Module' attribute was added to the manifest file in the 'jar' task in the build.gradle file. The module name was set as 'eu.lilithmonodia.winestock', ensuring that the main class and all its dependencies are encapsulated in this module when the jar file is created.