"Add redirection rule for 'How-to' page in WineStock
A new redirection rule was added in the 'redirection-rules.xml' document. This was created due to the removal of the 'How-to' page from WineStock. Now, whenever a user tries to access the 'How-to.html' page, they will be redirected according to the new rule. This change helps maintain the website's user experience despite the removal of the 'How-to' page."