Update Assortment class to be more flexible
The Assortment class has been updated to use generics, which should increase its reusability and reduce the need for extensive changes in the case that new product types need to be included in an assortment in the future. This not only makes it more flexible, but reduces the possibility of code duplication. This update has also been reflected in the relevant classes that use the Assortment class.