Refactor Assortment class and enhance unit tests
This commit majorly has refactoring of Assortment class which updates the TableView from simply viewing Assortments to view Assortments of Wines, thus making it more specific and accurate. This change is dysplayed in the refactor of TableColumn in Assortment. Also, relevant changes were made in the AssortmentTest to align with the refactoring done in Assortment. This not only ensures thorough testing of the assortment class features like add, remove, retain and sort but also improves code readability and efficiency.