Adjust testRemoveObject in AssortmentTest class
Method testRemoveObject in AssortmentTest class was modified to add the instance of wine before removing it. This way, the test case is more solid and exact because it ensures the wine instance was properly added before attempting to remove and validate the removal. The same has been tracked in the main Assortment class by elaborating its method documentation for better readability.