Add 'wine.svg' to the docs/images directory
Added a new SVG image, 'wine.svg', to the project in the docs/images directory. This image will be used in the documentation to illustrate the concept of the 'Wine' feature of the project. The addition of this image aims to enhance the user's understanding of the feature, making the documentation more visually appealing and intuitive.