Update version number in fxml file and gradle build file
Updated the version number of the application from 0.6.4 to 0.6.5 in the WineStock.fxml and build.gradle files respectively. The reason for this alteration is to keep the version number up-to-date following recent modifications to the application. This change is important to ensure that users and developers are working with and are aware of the correct version of the application.