Refactor toString() method in Assortment class
The diff portrays a modification in the toString() method of the Assortment class. Earlier, method was returning the string representation of the 'wineList' directly. Now, it's been refactored to use a StringBuilder for generating a more detailed, formatted string representation of the Assortment object - including information of 'id', 'wines' and 'totalPrice'. This enhances understandability of the object's info while debugging or logging.