Add port field to DB connection form and controller
Added a new TextField input for specifying the port in the database connection form. The host field has been embedded in a HBox with the new port field. The default value (5432) of port will be filled if the port field is empty. This allows users to connect to PostgreSQL databases running on non-default ports, increasing the flexibility of the app.