Add array of TableColumn fields for assortment table
Added an array of TableColumn fields to handle the properties of each wine in the assortment table. This enables the user to easily manage different properties of wines in the assortment through the user interface. Refactoring was also performed to reduce duplicate code with value factories across different tables. New functionality was added to load selected assortments into the table. The organization and clarity of the code have been significantly improved.