"Remove deleteAssortmentInternal method from PostgreSQLManager
Deleted the private method deleteAssortmentInternal in PostgreSQLManager.java. This method was previously responsible for deleting an assortment from the database. The functionality of this method has been integrated with maintenance procedures handled by the WineStockController as now we aim to add and removed wines directly from assortments, instead of deleting the whole assortment. The change simplifies the codebase and improves maintainability."