Update application version to 1.1.1 in various files
Updated the application version from 1.1.0 to 1.1.1 in build.gradle, WineStock.fxml and README.md files. The changes are made in order to align the mentioned version number with the latest application updates and releases. Updated version in documentation and source code fosters better version tracking and clarification about application state.