Update 'WineStock.fxml' to enhance visuals
The given diff represents major enhancements in the 'WineStock.fxml' file which improves the outlook and user experience of the GUI. It includes updates like the import of javafx properties that enhance the components of the GUI, updates on different tabs like 'Connexion', 'Vins', 'Assortiments', and 'À Propos'. The dimensions, maximum and minimum heights, widths, etc., of multiple graphical components have been altered to create a better user interface. It also includes the addition and deletion of some VBox, Buttons, HBox, TextField, etc.