Updated JavaFX namespace in WineStock.fxml
In this commit, the namespace of JavaFX in WineStock.fxml has been depreciated from '20.0.1' to '17.0.0'. This is in line with the project's decision to support a lower version of JavaFX due to compatibility issues reported. Code formatting has also been adjusted for improved readability, comprising mainly of lines shifted resulting in no change in logic or functionality.