Optimize import statements in WineStock.fxml
Restructured import statements, removed unnecessary imports and grouped similar imports together in WineStock.fxml file. The layout of the elements were also updated, making it look more clear, concise and easy to read. Code logic is not changed.

Also, made changes to the GUI prompts to  support both French and English language users more effectively. For example, previous prompt text "Wine Name" is updated to "Nom" (which means 'name' in French). It is part of an ongoing effort to internationalize the app.