Set default values for wineVolumeComboBox and wineColorComboBox
Enhanced the wine create functionality. Initialized the 'wineVolumeComboBox' to 'BOUTEILLE' and 'wineColorComboBox' to 'ROUGE' as default values. This simplifies the user input process in scenarios where the user does not have a specific preference. Also updated the deleteWine method with a check to ensure a Wine object is selected and a proper error message is shown. This prevents crashes when a delete operation is attempted without a selected Wine object.