Adjusted deleteWineFromAssortment logic to improve error handling
A more robust error handling has been introduced to the deleteWinesFromAssortment method in the WineStockController.java file. Additional checks were implemented for if the currentlySelectedAssortment is null and if no wine has been selected to prevent unhandled exceptions. This change simplifies error handling by using a single function for displaying app errors, which results in cleaner and more maintainable code.