Update UI elements for better clarity in WineStock app
This commit bumps the Winestock version from 1.1.2 to 1.1.3 and introduces various changes in the 'About' Tab for better user-friendliness. Interface-related imports and attributes have been updated and restructured in 'WineStock.fxml' file to facilitate better clarity and readability for the end-user. This involves revising the layout for Wine tables, adding refresh buttons and minor text changes. The UI changes help to improve the UX and promote better navigation within the application.