Add new library Atlantafx-base and UI tweaks
Introduce the Atlantafx-base library in gradle file for better UI/JavaFX controls and layouts. Make multiple changes in the FXML to tweak the UI. Changes include better alignment, increased margins, better widths for columns, and more organized imports. These changes aim to improve the look and feel of the application.