Add style classes to various elements in WineStock.fxml
A new style has been applied to several TableView elements and in the TextArea section in our WineStock.fxml. The style classes "stripped" and "dense" have been added to the TableView elements to enhance the visual experience of how data is displayed. The TextArea element has been enhanced with rounded style for better look when rendering the MIT license's content. These changes intend to provide a more visually appealing and user-friendly application interface.