Update project version to 1.2.0 and modify javafx version in fxml file
The project version has been updated from '1.1.3' to '1.2.0' in build.gradle file denoting the release of new version of the application. The javafx version in WineStock.fxml file is also changed from '20.0.1' to '17.0.7' to ensure compatibility with the current environment. Apart from these, the displayed version in 'About' tab of the application GUI has also been updated corresponding to the project version.