Update UML diagram in README
The UML class diagram link in the README file has been updated from 'uml-diagram-1.1.1.png' to 'uml-diagram-1.2.0.png'. This change reflects the updated structure of the application in the latest version 1.2.0. The README file should always represent the up-to-date state of the project, so it’s necessary to replace old diagrams with the new ones after significant architectural changes.