Removed style classes on database import start
Adjusted the importDatabase() method in WineStockController to remove previous styling classes from the importButton when a new import process is initiated. This change ensures that any success or danger styles from previous import actions are cleared, providing a clean slate for the next import feedback. This leads to better UX by providing accurate and timely success/error feedback to the user.