Add @FXML annotation to refresh method
The refresh method in the WineStockController class is now annotated with @FXML. This annotation is necessary because it tells JavaFX to inject an object where a value from the markup can be injected, which will help to connect the model view (TableView) with the controller object (WineStockController). The refresh method replaces TableView items with data from PostgreSQL database.