Update GitHub workflows and Gradle settings for jpackage
The adjustments change the build workflow to use chmod and Gradle with jpackage, and include an update to the installer options in the build script. The purpose of this is to enable the building of different packages for different OS types. Also updated upload configurations in the GitHub workflow to accommodate the new types of packages being built.